some small refactoring to make main smaller and easier to read.
Notify on entry moved to utils.py

check_ip() moved to utils.py
